Possible optimization for low-end systems?

+++ src/bpp_shared/world/noise/noise_simplex.cpp
@@ -43,24 +43,24 @@ void NoiseSimplex::GenerateNoise(std::span<double> _noiseField, Vec2 _offset, In
                                  double _amplitude) {
 	double* out = _noiseField.data();
 
-	std::vector<double> yPositions(_size.y);
+	std::vector<float> yPositions(_size.y);
 	for (int32_t yI = 0; yI < _size.y; ++yI) {
-		yPositions[yI] = (_offset.y + double(yI)) * _scale.y + coordinate.y;
+		yPositions[yI] = (_offset.y + float(yI)) * _scale.y + coordinate.y;
 	}
 
 	for (int32_t xI = 0; xI < _size.x; ++xI) {
-		double xPos = (_offset.x + double(xI)) * _scale.x + coordinate.x;
+		float xPos = (_offset.x + float(xI)) * _scale.x + coordinate.x;
 
 		for (int32_t yI = 0; yI < _size.y; ++yI) {
-			double yPos = yPositions[yI];
-			double skew = (xPos + yPos) * SKEWING;
+			float yPos = yPositions[yI];
+			float skew = (xPos + yPos) * SKEWING;
 			int32_t x0 = Wrap(xPos + skew);
 			int32_t y0 = Wrap(yPos + skew);
-			double unskewed = double(x0 + y0) * UNSKEWING;
-			double x0a = double(x0) - unskewed;
-			double y0a = double(y0) - unskewed;
-			double x0b = xPos - x0a;
-			double y0b = yPos - y0a;
+			float unskewed = float(x0 + y0) * UNSKEWING;
+			float x0a = float(x0) - unskewed;
+			float y0a = float(y0) - unskewed;
+			float x0b = xPos - x0a;
+			float y0b = yPos - y0a;
 			int8_t i;
 			int8_t j;
 			if (x0b > y0b) {
@@ -71,17 +71,17 @@ void NoiseSimplex::GenerateNoise(std::span<double> _noiseField, Vec2 _offset, In
 				j = 1;
 			}
 
-			double x0c = x0b - double(i) + UNSKEWING;
-			double y0c = y0b - double(j) + UNSKEWING;
-			double x1c = x0b - 1.0 + 2.0 * UNSKEWING;
-			double y1c = y0b - 1.0 + 2.0 * UNSKEWING;
+			float x0c = x0b - float(i) + UNSKEWING;
+			float y0c = y0b - float(j) + UNSKEWING;
+			float x1c = x0b - 1.0 + 2.0 * UNSKEWING;
+			float y1c = y0b - 1.0 + 2.0 * UNSKEWING;
 			int32_t xInt = x0 & 255;
 			int32_t yInt = y0 & 255;
 			int32_t grad0 = permMod12[xInt + permutations[yInt]];
 			int32_t grad1 = permMod12[xInt + i + permutations[yInt + j]];
 			int32_t grad2 = permMod12[xInt + 1 + permutations[yInt + 1]];
-			double term0 = 0.5 - x0b * x0b - y0b * y0b;
-			double contrib0;
+			float term0 = 0.5 - x0b * x0b - y0b * y0b;
+			float contrib0;
 			if (term0 < 0.0) {
 				contrib0 = 0.0;
 			} else {
@@ -89,8 +89,8 @@ void NoiseSimplex::GenerateNoise(std::span<double> _noiseField, Vec2 _offset, In
 				contrib0 = term0 * term0 * DotProd(GRADIENTS[grad0], x0b, y0b);
 			}
 
-			double term1 = 0.5 - x0c * x0c - y0c * y0c;
-			double contrib1;
+			float term1 = 0.5 - x0c * x0c - y0c * y0c;
+			float contrib1;
 			if (term1 < 0.0) {
 				contrib1 = 0.0;
 			} else {
@@ -98,8 +98,8 @@ void NoiseSimplex::GenerateNoise(std::span<double> _noiseField, Vec2 _offset, In
 				contrib1 = term1 * term1 * DotProd(GRADIENTS[grad1], x0c, y0c);
 			}
 
-			double term2 = 0.5 - x1c * x1c - y1c * y1c;
-			double contrib2;
+			float term2 = 0.5 - x1c * x1c - y1c * y1c;
+			float contrib2;
 			if (term2 < 0.0) {
 				contrib2 = 0.0;
 			} else {
@@ -107,7 +107,7 @@ void NoiseSimplex::GenerateNoise(std::span<double> _noiseField, Vec2 _offset, In
 				contrib2 = term2 * term2 * DotProd(GRADIENTS[grad2], x1c, y1c);
 			}
 
-			*out++ += 70.0 * (contrib0 + contrib1 + contrib2) * _amplitude;
+			*out++ += double(70.0 * (contrib0 + contrib1 + contrib2) * _amplitude);
 		}
 	}
 }
